♻️ [Refactoring]: #221 list wire を commands 層へ分離し application 層から serde を除去
spec 046 ステップ 2 本体:

- src/commands/list.rs を module root 化し wire/json/table/simple/outdated の
  フォーマット別サブモジュールへ分割
- src/commands/list/wire.rs に Wire / ComponentsWire / OutdatedWire を新設
  （marketplace は skip_serializing_if で現状互換を維持）
- outdated は render_outdated_json（HostClient 非依存の純粋 render）と
  print_outdated_* / run_outdated（live 実行）を分離配置
- JSON / outdated の byte-for-byte 互換性をスナップショットテストで保証
- src/application/plugin_catalog.rs から InstalledPlugin / Serialize impl /
  ComponentsByKind を削除し、list_installed_plugins は InstalledPlugin::
  from_cached_package を呼ぶ形に置換
- src/application/plugin_info.rs の InstalledPlugin::new を from_cached_package に置換
- src/application.rs の pub use を plugin_catalog::list_installed_plugins +
  crate::plugin::InstalledPlugin に差し替え（TUI 側 import を再エクスポートで温存）
- plugin_catalog_test.rs のアクセサ系テストを installed_test.rs へ分離

これにより src/application/ 配下から serde 属性・impl・use が完全に消え、
application レイヤーは pure use case 関数のみを持つ構成になる。
