+++ src/application.rs
@@ -10,8 +10,9 @@ mod plugin_info;
 mod plugin_intent;
 mod plugin_operations;
 
+pub use crate::plugin::InstalledPlugin;
 pub use crate::target::OperationResult;
-pub use plugin_catalog::{list_installed_plugins, InstalledPlugin};
+pub use plugin_catalog::list_installed_plugins;
 pub use plugin_info::{get_plugin_info, PluginInfo, Source};
 pub use plugin_operations::{
     disable_plugin, enable_plugin, get_uninstall_info, uninstall_plugin, UninstallInfo,
